Code cleaning and comments

+++ Entities/Platform.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 Platform::Platform(Scene *scene, const sf::Texture &texture) : GraphicalEntity( scene, "Platform", texture) {
-    setPosition( (scene->getWindowWidth()-this->getWidth())/2.0, scene->getWindowHeight() - this->getHeight() - 10);
+    setPosition((float) ((scene->getWindowWidth() - this->getWidth()) / 2.0), scene->getWindowHeight() - this->getHeight() - 10);
     velocity = 6.0;
 }
